+++ docs-src/pages/UiSelect.vue
@@ -333,6 +333,17 @@
                             </tr>
 
                             <tr>
+                                <td>sort</td>
+                                <td>Function</td>
+                                <td></td>
+                                <td>
+                                    <p>Defines a custom sort function that is used for sorting the options after they are filtered.</p>
+
+                                    <p>The function should implement the standard <a href="https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/sort#Syntax" target="_blank" rel="noopener"><code>Array.prototype.sort()</code> compare callback</a>.</p>
+                                </td>
+                            </tr>
+
+                            <tr>
                                 <td>loading</td>
                                 <td>Boolean</td>
                                 <td><code>false</code></td>